cleanup build warnings, comments etc.
