+++ chain/src/chain.rs
@@ -446,6 +446,7 @@ impl Chain {
 		}
 	}
 
+	/// Validate the tx against the current UTXO set.
 	pub fn validate_tx(&self, tx: &Transaction) -> Result<(), Error> {
 		let mut txhashset = self.txhashset.write().unwrap();
 		txhashset::utxo_view(&mut txhashset, |utxo| {
