+++ core/src/core/pmmr.rs
@@ -44,6 +44,7 @@ use croaring::Bitmap;
 use core::hash::Hash;
 use core::merkle_proof::MerkleProof;
 use core::BlockHeader;
+use core::readonly_pmmr::ReadonlyPMMR;
 use ser::{PMMRIndexHashable, PMMRable};
 use util::LOGGER;
 
@@ -125,57 +126,6 @@ where
 	_marker: marker::PhantomData<T>,
 }
 
-pub struct PMMRReadonly<'a, T, B>
-where
-	T: PMMRable,
-	B: 'a + Backend<T>,
-{
-	/// The last position in the PMMR
-	last_pos: u64,
-	/// The backend for this readonly PMMR
-	backend: &'a B,
-	// only needed to parameterise Backend
-	_marker: marker::PhantomData<T>,
-}
-
-impl<'a, T, B> PMMRReadonly<'a, T, B>
-where
-	T: PMMRable + ::std::fmt::Debug,
-	B: 'a + Backend<T>,
-{
-	pub fn new(backend: &'a B) -> PMMRReadonly<T, B> {
-		PMMRReadonly {
-			last_pos: 0,
-			backend: backend,
-			_marker: marker::PhantomData,
-		}
-	}
-
-	/// Build a new readonly PMMR pre-initialized to
-	/// last_pos with the provided backend.
-	pub fn at(backend: &'a B, last_pos: u64) -> PMMRReadonly<T, B> {
-		PMMRReadonly {
-			last_pos: last_pos,
-			backend: backend,
-			_marker: marker::PhantomData,
-		}
-	}
-
-	/// Get the data element at provided position in the MMR.
-	pub fn get_data(&self, pos: u64) -> Option<T> {
-		if pos > self.last_pos {
-			// If we are beyond the rhs of the MMR return None.
-			None
-		} else if is_leaf(pos) {
-			// If we are a leaf then get data from the backend.
-			self.backend.get_data(pos)
-		} else {
-			// If we are not a leaf then return None as only leaves have data.
-			None
-		}
-	}
-}
-
 impl<'a, T, B> PMMR<'a, T, B>
 where
 	T: PMMRable + ::std::fmt::Debug,
@@ -200,8 +150,9 @@ where
 		}
 	}
 
-	pub fn pmmr_readonly(&self) -> PMMRReadonly<T, B> {
-		PMMRReadonly::at(&self.backend, self.last_pos)
+	/// Build a "readonly" view of this PMMR.
+	pub fn readonly_pmmr(&self) -> ReadonlyPMMR<T, B> {
+		ReadonlyPMMR::at(&self.backend, self.last_pos)
 	}
 
 	/// Returns a vec of the peaks of this MMR.
